requirements.txt
st-pages==0.4.5
streamlit==1.27.2
streamlit-feedback==0.1.2
requests
pyyaml
pypdf2
openai>=1.32.0,<2.0.0google-generativeai

tqdm
tiktoken
psycopg2-binary
sqlalchemy

pymysql
streamlit
altair
geopy

unidecode
cryptography

python-docx
python-pptx
beautifulsoup4